doice/requirements.txt
django
djangorestframework
pillow
django-rest-auth
django-betterforms
django-rest-knox